Implemented the 'reset vehicle' features into the main scenes

+++ Assets/Micro Motors Assets/Scripts/Player/GlobalCash.cs
@@ -11,7 +11,7 @@ public class GlobalCash : MonoBehaviour
     
     void Start()
     {
-        TotalCredits = PlayerPrefs.GetInt("SavedCash");
+        TotalCredits = PlayerPrefs.GetFloat("SavedCash");
     }
 
     void Update()
